Fix service module filenames for consistency and move common code to separate file
